# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
...
paths:
...
  /api/github/issues:
    get:
      operationId: listGitHubIssues
      summary: List GitHub issues for a workspace codebase
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: string
        - name: codebaseId
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed, all]
      responses:
        "200":
          description: GitHub issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  repo:
                    type: string
                  codebase:
                    type: object
                    properties:
                      id:
                        type: string
                      label:
                        type: string
                  issues:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        number:
                          type: integer
                        title:
                          type: string
                        body:
                          type: string
                        url:
                          type: string
                        state:
                          type: string
                        labels:
                          type: array
                          items:
                            type: string
                        assignees:
                          type: array
                          items:
                            type: string
                        updatedAt:
                          type: string
        "400":
          description: Missing workspace or non-GitHub codebase
        "404":
          description: Workspace codebase not found
        "502":
          description: GitHub upstream request failed
...
  /api/workflows:
    get:
      operationId: listWorkflows
      summary: List all workflow YAML definitions from resources/flows/
      responses:
        "200":
          description: Workflow list
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        version:
                          type: string
                        trigger:
                          type: object
                        steps:
                          type: array
                        yamlContent:
                          type: string
    post:
      operationId: createWorkflow
      summary: Create a new workflow YAML file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, yamlContent]
              properties:
                id:
                  type: string
                yamlContent:
                  type: string
      responses:
        "201":
          description: Workflow created
        "400":
          description: Invalid input
        "409":
          description: Workflow already exists
  /api/workflows/{id}:
    get:
      operationId: getWorkflow
      summary: Get a specific workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow detail
        "404":
          description: Workflow not found
    put:
      operationId: updateWorkflow
      summary: Update a workflow YAML file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [yamlContent]
              properties:
                yamlContent:
                  type: string
      responses:
        "200":
          description: Workflow updated
        "400":
          description: Invalid input
        "404":
          description: Workflow not found
    delete:
      operationId: deleteWorkflow
      summary: Delete a workflow YAML file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow deleted
        "404":
          description: Workflow not found
  /api/workflows/{id}/trigger:
    post:
      operationId: triggerWorkflow
      summary: Trigger a workflow run inside a workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId]
              properties:
                workspaceId:
                  type: string
                triggerPayload:
                  type: string
      responses:
        "201":
          description: Workflow run created
          content:
            application/json:
              schema:
                type: object
                required: [workflowRunId, taskIds]
                properties:
                  workflowRunId:
                    type: string
                  taskIds:
                    type: array
                    items:
                      type: string
        "400":
          description: Invalid input
        "404":
          description: Workflow or workspace not found

  # ── Debug (Rust desktop only) ──
...